+++ CHANGELOG.md
@@ -51,6 +51,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Show the project's Sponsor trust required information in the project
   information tab
 - New Process the Sponsored support grant task added to the Voluntary task list
+- when viewing a completed project, a banner is shown to inform users why they
+  cannot change the project
 
 ### Fixed